Stop delta updates at old repetitions as well
Add a sticky flag that holds the "was ever handled as a repetition" information
for edge. Values sent from such nodes can are a mix of normal values and draws
and are hard to adjust using the delta update. So stop delta updates at such
edges every time, both when they are handled as a repetition in the current
backup or were handled as a repetition during some previous backup.